Switch the docs to the "Furo" theme
Pros:
(+) It looks cleaner.
(+) It utilises all the width (or more width).
(+) It has dark/light modes for those who care.
(+) It separates the main ToC on the left from the intra-page navigation on the right.
(+) Pytest uses it, so keeping convenient within the domain of Python testing.

Cons: None.

The library is fresh, so there will be no struggles reading the docs either way.

Signed-off-by: Sergey Vasilyev <nolar@nolar.info>